modify update function

+++ app/Http/Controllers/ArticleController.php
@@ -86,11 +86,13 @@ class ArticleController extends Controller
         $article->title = $request->title;
         $article->content = $request->content;
 
+       if($request->hasFile('image')){
         if ($article->featured_image && file_exists(storage_path('app/public/'.$article->featured_image))) {
             Storage::delete('public/'.$article->featured_image);
         }
         $image_name = $request->file('image')->store('images','public');
         $article->featured_image = $image_name;
+       }
 
         $article->save();
         return 'Artikel berhasil diubah';
